events: Rename enums module to common
