+++ crates/ruma-macros/src/events.rs
@@ -1,7 +1,7 @@
 //! Methods and types for generating events.
 
-pub mod enums;
-pub mod event;
-pub mod event_content;
-pub mod event_enum;
-pub mod event_enum_from_event;
+pub(crate) mod common;
+pub(crate) mod event;
+pub(crate) mod event_content;
+pub(crate) mod event_enum;
+pub(crate) mod event_enum_from_event;
