+++ crates/ruma-macros/src/events/common.rs
