+++ crates/ruma-macros/src/events/event_content/parse.rs
@@ -10,7 +10,7 @@ use syn::{
 
 use super::{EventContent, EventContentField, EventContentKind};
 use crate::{
-    events::enums::{EventKind, EventType, EventTypes},
+    events::common::{EventKind, EventType, EventTypes},
     util::{ParseNestedMetaExt, RumaEvents, SerdeMetaItem, StructFieldExt},
 };
